Fix simultaneous repair/death animation
When a bot is "killed" on the same turn that it is repairing (meaning it does not die, because the repair saves it) both the death and repair animations were being shown. This change moves the check for dead bots outside of the bot action executor loop so we only show a death animation if a bot actually dies.

Note that this is still not the best way to fix this problem, as we're just putting all death animations onto the first (possibly-unrelated) action result. However, this works for now until we have time to refactor more in the future.